Separate PalantirDockerPlugin and DockerComposeTemplatePlugin into two plugins
