+++ readme.md
@@ -2,12 +2,19 @@ Docker Gradle Plugin
 ====================
 [![Build Status](https://travis-ci.org/palantir/gradle-docker.svg?branch=develop)](https://travis-ci.org/palantir/gradle-docker)
 
-Adds basic tasks for building and pushing docker images based on a simple
-configuration block that specifies the container name, the Dockerfile, task
-dependencies, and any additional file resources required for the Docker build.
+This repository provides two Gradle plugins for working with Docker containers:
+- The `com.palantir.docker` plugin add basic tasks for building and pushing
+docker images based on a simple configuration block that specifies the container
+name, the Dockerfile, task dependencies, and any additional file resources
+required for the Docker build.
+- The `com.palantir.docker-compose` plugin adds a task for populating
+placeholders in a docker-compose template file with image versions resolved from
+dependencies.
 
-Usage
------
+Docker Plugin
+-------------
+
+### Usage
 Apply the plugin using standard gradle convention:
 
     plugins {
@@ -33,34 +40,35 @@ Configuration specifying all parameters:
         dockerfile 'Dockerfile'
         dependsOn tasks.distTar
         files 'file1.txt', 'file2.txt'
-        dockerComposeTemplate 'my-template.yml'
-        dockerComposeFile 'my-docker-compose.yml'
     }
 
 To build a docker container, run the `docker` task. To push that container to a
 docker repository, run the `dockerPush` task.
 
 
-Generating docker-compose dependencies
---------------------------------------
+Managing Docker image dependencies
+----------------------------------
 
-The plugin provides mechanisms for managing dependencies between docker images
-used in orchestrating `docker-compose` environments. Each project can declare
-its dependencies on other docker images and publish an artifact advertising those
-dependencies. The plugin uses standard Maven/Ivy machanism for declaring and
-resolving dependencies.
+The `com.palantir.docker` and `com.palantir.docker-compose` plugins provide
+functionality to declare and resolve version-aware dependencies between docker
+images. The primary use-case is to generate `docker-compose.yml` files whose
+image versions are mutually compatible and up-to-date in cases where multiple
+images depend on the existence of the same Dockerized service.
 
-The `generateDockerCompose` task generates a `docker-compose.yml` file from
-a user-defined template by replacing each version variable by the concrete version
-declared by the transitive dependencies of the docker configuration.
+### Specifying and publishing dependencies on Docker images
 
-### Specifying dependencies on Docker images
-The plugin adds a `docker` Gradle component and a `docker` Gradle configuration that
-can be used to specify and publish dependencies on other Docker containers.
-The `generateDockerCompose` task (see below) uses the transitive dependencies
-of the `docker` configuration to populate a `docker-compose.yml.template` file
-with the image versions specified by this project and all its transitive
-dependencies.
+The `docker` plugin adds a `docker` Gradle component and a `docker` Gradle
+configuration that can be used to specify and publish dependencies on other
+Docker containers.
+
+**Example**
+
+    plugins {
+        id 'maven-publish'
+        id 'com.palantir.docker'
+    }
+
+    ...
 
     dependencies {
         docker 'foogroup:barmodule:0.1.2'
@@ -76,16 +84,28 @@ dependencies.
         }
     }
 
-The above configuration adds a Maven publication that specifies dependencies
-on `barmodule` and the `someSubProject` Gradle sub project. The resulting POM
-file has two `dependency` entries, one for each dependency.
+The above configuration adds a Maven publication that specifies dependencies on
+`barmodule` and the `someSubProject` Gradle sub project. The resulting POM file
+has two `dependency` entries, one for each dependency. Each project can declare
+its dependencies on other docker images and publish an artifact advertising
+those dependencies.
 
-### Generating docker-compose.yml files from templates
-The `generateDockerCompose` task performs two operations: First, it generates
-a mapping `group:name --> version` from the dependencies of the `docker`
-configuration (see above). Second, it replaces all occurrences of version
-variables of the form `{{group:name}}` in the `docker-compose.yml.template` file
-by the resolved versions and writes the resulting file as `docker-compose.yml`.
+### Generating docker-compose.yml files from dependencies
+
+The `com.palantir.docker-compose` plugin uses the transitive dependencies of the
+`docker` configuration to populate a `docker-compose.yml.template` file with the
+image versions specified by this project and all its transitive dependencies.
+The plugin uses standard Maven/Ivy machanism for declaring and resolving
+dependencies.
+
+The `generateDockerCompose` task generates a `docker-compose.yml` file from a
+user-defined template by replacing each version variable by the concrete version
+declared by the transitive dependencies of the docker configuration.  The task
+performs two operations: First, it generates a mapping `group:name --> version`
+from the dependencies of the `docker` configuration (see above). Second, it
+replaces all occurrences of version variables of the form `{{group:name}}` in
+the `docker-compose.yml.template` file by the resolved versions and writes the
+resulting file as `docker-compose.yml`.
 
 **Example**
 
@@ -96,34 +116,15 @@ Assume a `docker-compose.yml.template` as follows:
     otherservice:
       image: 'repository/otherservice:{{othergroup:otherservice}}'
 
-The `build.gradle` of this project publishes a docker Maven artifact declaring
-a dependency on a docker image published as 'othergroup:otherservice' in version
-0.1.2:
+`build.gradle` declares a dependency on a docker image published as
+'othergroup:otherservice' in version 0.1.2:
 
     plugins {
-        id 'maven-publish'
-        id 'com.palantir.docker'
+        id 'com.palantir.docker-compose'
     }
-
-    docker {
-        name 'foo'
-    }
-
-    group 'mygroup'
-    name 'myservice'
-    version '2.3.4'
 
     dependencies {
         docker 'othergroup:otherservice:0.1.2'
-    }
-
-    publishing {
-        publications {
-            dockerPublication(MavenPublication) {
-                from components.docker
-                artifactId project.name + "-docker"
-            }
-        }
     }
 
 The `generateDockerCompose` task creates a `docker-compose.yml` as follows:
@@ -133,10 +134,21 @@ The `generateDockerCompose` task creates a `docker-compose.yml` as follows:
     otherservice:
       image: 'repository/otherservice:0.1.2'
 
-The `generateDockerCompose` task fails if the template file contains variables that
-cannot get resolved using the provided `docker` dependencies. Version conflicts between
-transitive dependencies of the same artifact are handled with the standard Gradle
-semantics: each artifact is resolved to the highest declared version.
+The `generateDockerCompose` task fails if the template file contains variables
+that cannot get resolved using the provided `docker` dependencies. Version
+conflicts between transitive dependencies of the same artifact are handled with
+the standard Gradle semantics: each artifact is resolved to the highest declared
+version.
+
+**Configuring file locations**
+
+The template and generated file locations are customizable through the
+`dockerCompose` extension:
+
+    dockerCompose {
+        template 'my-template.yml'
+        dockerComposeFile 'my-docker-compose.yml'
+    }
 
 Tasks
 -----
@@ -150,6 +162,8 @@ Tasks
  * `dockerfileZip`: builds a ZIP file containing the configured Dockerfile
  * `generateDockerCompose`: Populates a docker-compose file template with image
    versions declared by dependencies
+
+
