+++ src/main/resources/META-INF/gradle-plugins/com.palantir.docker-compose.properties
@@ -1,0 +1,1 @@
+implementation-class=com.palantir.gradle.docker.DockerComposePlugin
