+++ src/test/groovy/com/palantir/gradle/docker/AbstractPluginTest.groovy
@@ -1,0 +1,45 @@
+package com.palantir.gradle.docker
+
+import org.gradle.testkit.runner.GradleRunner
+import org.junit.Rule
+import org.junit.rules.TemporaryFolder
+import spock.lang.Specification
+
+class AbstractPluginTest extends Specification {
+
+    @Rule
+    TemporaryFolder temporaryFolder = new TemporaryFolder()
+
+    File projectDir
+    File buildFile
+    List<File> pluginClasspath
+
+    GradleRunner with(String... tasks) {
+        GradleRunner.create()
+            .withPluginClasspath(pluginClasspath)
+            .withProjectDir(projectDir)
+            .withArguments(tasks)
+    }
+
+    String exec(String task) {
+        StringBuffer sout = new StringBuffer(), serr = new StringBuffer()
+        Process proc = task.execute()
+        proc.consumeProcessOutput(sout, serr)
+        proc.waitFor()
+        return sout.toString()
+    }
+
+    def setup() {
+        projectDir = temporaryFolder.root
+        buildFile = temporaryFolder.newFile('build.gradle')
+
+        def pluginClasspathResource = getClass().classLoader.findResource("plugin-classpath.txt")
+        if (pluginClasspathResource == null) {
+            throw new IllegalStateException("Did not find plugin classpath resource, run `testClasses` build task.")
+        }
+
+        pluginClasspath = pluginClasspathResource.readLines()
+            .collect { it.replace('\\', '\\\\') } // escape backslashes in Windows paths
+            .collect { new File(it) }
+    }
+}
