+++ src/test/groovy/com/palantir/gradle/docker/PalantirDockerPluginTests.groovy
@@ -19,22 +19,9 @@ import org.gradle.api.internal.artifacts.mvnsettings.DefaultLocalMavenRepository
 import org.gradle.api.internal.artifacts.mvnsettings.DefaultMavenFileLocations
 import org.gradle.api.internal.artifacts.mvnsettings.DefaultMavenSettingsProvider
 import org.gradle.testkit.runner.BuildResult
-import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
-import org.junit.Rule
-import org.junit.rules.TemporaryFolder
 
-import spock.lang.Specification
-
-class PalantirDockerPluginTests extends Specification {
-
-    @Rule
-    TemporaryFolder temporaryFolder = new TemporaryFolder()
-
-    File projectDir
-    File buildFile
-    List<File> pluginClasspath
-
+class PalantirDockerPluginTests extends AbstractPluginTest {
 
     def 'fail when missing docker configuration'() {
         given:
@@ -238,155 +225,4 @@ class PalantirDockerPluginTests extends Specification {
         ["foogroup", "barmodule", "0.1.2", temporaryFolder.root.name, "2.3.4"].each { dockerPom.contains(it) }
         !dockerPom.contains("guava")
     }
-
-    def 'Generates docker-compose.yml from template with version strings replaced'() {
-        given:
-        temporaryFolder.newFile('Dockerfile') << "Foo"
-        temporaryFolder.newFile("docker-compose.yml.template") << '''
-            service1:
-              image: 'repository/service1:{{com.google.guava:guava}}'
-            service2:
-              image: 'repository/service2:{{org.slf4j:slf4j-api}}'
-        '''.stripIndent()
-        buildFile << '''
-            plugins {
-                id 'com.palantir.docker'
-            }
-
-            repositories {
-                jcenter()
-            }
-
-            docker {
-                name 'foo'
-            }
-
-            dependencies {
-                docker 'io.dropwizard:dropwizard-jackson:0.8.2'
-                  // transitive dependencies: com.google.guava:guava:18.0, org.slf4j:slf4j-api:1.7.10
-                docker 'com.google.guava:guava:17.0'  // should bump to 18.0 via the above
-            }
-
-        '''.stripIndent()
-
-        when:
-        BuildResult buildResult = with('generateDockerCompose', "--stacktrace").build()
-        then:
-        buildResult.task(':generateDockerCompose').outcome == TaskOutcome.SUCCESS
-        def dockerComposeText = temporaryFolder.root.toPath().resolve("docker-compose.yml").text
-        dockerComposeText.contains("repository/service1:18.0")
-        dockerComposeText.contains("repository/service2:1.7.10")
-    }
-
-    def 'Fails if docker-compose.yml.template has unmatched version tokens'() {
-        given:
-        temporaryFolder.newFile('Dockerfile') << "Foo"
-        temporaryFolder.newFile("docker-compose.yml.template") << '''
-            service1:
-              image: 'repository/service1:{{foo:bar}}'
-        '''.stripIndent()
-        buildFile << '''
-            plugins {
-                id 'com.palantir.docker'
-            }
-
-            repositories {
-                jcenter()
-            }
-
-            docker {
-                name 'foo'
-            }
-
-            dependencies {
-                docker 'com.google.guava:guava:17.0'  // should bump to 18.0 via the above
-            }
-
-        '''.stripIndent()
-
-        when:
-        BuildResult buildResult = with('generateDockerCompose', '--stacktrace').buildAndFail()
-        then:
-        buildResult.standardError.contains("Failed to resolve Docker dependencies mention in")
-        buildResult.standardError.contains("{{foo:bar}}")
-    }
-
-    def 'docker-compose template and file can have custom locations'() {
-        given:
-        temporaryFolder.newFile('Dockerfile') << "Foo"
-        new File(temporaryFolder.newFolder("templates"), "customTemplate.yml") << '''
-            nothing
-        '''.stripIndent()
-        buildFile << '''
-            plugins {
-                id 'com.palantir.docker'
-            }
-
-            repositories {
-                jcenter()
-            }
-
-            docker {
-                name 'foo'
-                dockerComposeTemplate 'templates/customTemplate.yml'
-                dockerComposeFile 'compose-files/customDockerCompose.yml'
-            }
-        '''.stripIndent()
-
-        when:
-        BuildResult buildResult = with('generateDockerCompose', "--stacktrace").build()
-        then:
-        buildResult.task(':generateDockerCompose').outcome == TaskOutcome.SUCCESS
-        temporaryFolder.root.toPath().resolve("compose-files").resolve("customDockerCompose.yml").toFile().exists()
-    }
-
-    def 'Fails if dockerComposeTemplate is configured but does not exist'() {
-        given:
-        temporaryFolder.newFile('Dockerfile') << "Foo"
-        buildFile << '''
-            plugins {
-                id 'com.palantir.docker'
-            }
-
-            docker {
-                name 'foo'
-                dockerComposeTemplate 'templates/customTemplate.yml'
-            }
-        '''.stripIndent()
-
-        when:
-        BuildResult buildResult = with('generateDockerCompose').buildAndFail()
-        then:
-        buildResult.standardError.contains("Could not find specified dockerComposeTemplate file")
-    }
-
-    private GradleRunner with(String... tasks) {
-        GradleRunner.create()
-            .withPluginClasspath(pluginClasspath)
-            .withProjectDir(projectDir)
-            .withArguments(tasks)
-    }
-
-    private String exec(String task) {
-        StringBuffer sout = new StringBuffer(), serr = new StringBuffer()
-        Process proc = task.execute()
-        proc.consumeProcessOutput(sout, serr)
-        proc.waitFor()
-        return sout.toString()
-    }
-
-    def setup() {
-        projectDir = temporaryFolder.root
-        buildFile = temporaryFolder.newFile('build.gradle')
-
-        def pluginClasspathResource = getClass().classLoader.findResource("plugin-classpath.txt")
-        if (pluginClasspathResource == null) {
-            throw new IllegalStateException("Did not find plugin classpath resource, run `testClasses` build task.")
-        }
-
-        pluginClasspath = pluginClasspathResource.readLines()
-            .collect { it.replace('\\', '\\\\') } // escape backslashes in Windows paths
-            .collect { new File(it) }
-    }
-
 }
